reorganised helper function structure into utils, fixed tests, redone the logic for 'check if deployed' to check for initialised authorisedGarages rather than marbles
